extract encoder helper methods into encoder_module
